Update memory_print API to be more extensible. Refactor implementations of the profile interpretation.
pretty_output.py: all functions for pretty printing placed here;
interpretations.py: interpretation functions or their callers placed here;
memory.py: improved autocalling to selected interpreation by getattr().
